rename: move prismSpy.test.ts -> prismaSpy.ts so jest doesn't run it
squash: fix import path
